+++ src/fluid_chan.h
@@ -57,8 +57,8 @@ int fluid_channel_pitch_bend(fluid_channel_t *chan, int val);
 int fluid_channel_pitch_wheel_sens(fluid_channel_t *chan, int val);
 int fluid_channel_get_cc(fluid_channel_t *chan, int num);
 int fluid_channel_get_num(fluid_channel_t *chan);
-void fluid_channel_set_interp_method(fluid_channel_t *chan, int new_method);
-int fluid_channel_get_interp_method(fluid_channel_t *chan);
+void fluid_channel_set_interp_method(fluid_channel_t *chan, uint8_t new_method);
+uint8_t fluid_channel_get_interp_method(fluid_channel_t *chan);
 
 #define fluid_channel_get_key_pressure(chan, key) ((chan)->key_pressure[key])
 #define fluid_channel_set_key_pressure(chan, key, val)                         \
